refactor: optimize publish(root situation)
1. parallel publish
2. merge commit and tag
3. build once
